pyproject.toml
[tool.poetry]
name = "macrostrat.embedding-tiler"
version = "0.1.0"
description = "A tile server that computes the relevance of polygons based on embedding text"
authors = ["Daven Quinn <code@davenquinn.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "macrostrat"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
mapbox-vector-tile = "^2.1.0"
macrostrat-utils = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
